[DEV] Add main function

+++ Yk_Typedef.h
@@ -13,7 +13,7 @@ typedef signed char           int8_t;
 typedef short              	  int16_t;
 typedef int                   int32_t;
 typedef float                 float32_t;
-typedef double                float32_t;
+typedef double                float64_t;
 typedef long long             int64_t;
 /*UnSigned*/
 typedef unsigned char         uint8_t;
